Fix FLEX decoding, front-end AFC, and DSP plumbing
FLEX decode fixes:
- Correct shift_rel handling in LTR/FLEX frame alignment
- Robust codeword vector walk through interleaved frame data
- Fix alpha (bw/decay) signature mismatch in demod path
- Slicer quality metric fixes for reliable 2/4-level decisions
- Parse numeric header fields correctly

Front-end fixes:
- Packet AFC: track burst center during AFSK reception (tests updated)
- Preserve AFC state across tune operations
- mon_gate fix in noise gate path
- Device.rate plumbing: consistent sample-rate accounting in radio device
- Gate AFC updates while locked to prevent drift
- Resampler uses i32 accumulator to avoid overflow

DSP plumbing:
- Capture mutex for safe buffer handoff
- Coalesce remix operations to avoid redundant work
- Image corrector for DC/imbalance artifacts
- NoiseFloor seeding on startup
- call_peak detection fix
- Remove SoftAgc (replaced by floor-based gain control)
- Cap autonotch correction depth
- Reject out-of-tolerance timing estimates

Frontend (web/index.html): updated UI with AFC/decode status surfaces.

+++ crates/engine/src/dmr/mod.rs
@@ -165,7 +165,9 @@ impl Resampler8k {
         let mut out = Vec::with_capacity(pcm.len() * 2);
         for &s in pcm {
             if self.started {
-                out.push(f32::from(self.prev + s) / 65536.0);
+                // i16 + i16 overflows at full-scale extremes (the vocoder
+                // can emit both samples near ±32767); widen before adding.
+                out.push((i32::from(self.prev) + i32::from(s)) as f32 / 65536.0);
             } else {
                 out.push(f32::from(s) / 32768.0);
                 self.started = true;
